refactor: 최단 경로 응답 객체를 적절한 path package로 이동
