+++ docs/step2-README.md
@@ -107,6 +107,7 @@ TDD를 기반으로 지하철 경로 조회 기능 구현
   - FetchType이 Lazy로 설정된 다른 엔티티와 연관관계를 가지는 경우 트랜잭션 밖에서 접근 시 `LazyInitializationException` 발생할 수 있음에 주의
   - 하위 계층인 `Persistence` 계층의 엔티티를 `Presentation` 계층의 DTO, Controller에 위치하는 것은 역할에 따라 계층을 분리하는 레이어드 아키텍쳐 관점에서 역방향으로 참조하고 있어 각 계층의 역할 경계를 흐트릴수 있고, 협업하는 다른 개발자로 하여금 혼선을 줄 여지가 있으므로 좋은 구조가 아님
 - [x] 코드 정합성 및 가독성을 위한 제네릭 형식 명시
+- [x] 최단 경로 응답 객체를 적절한 path package로 이동
 - [ ] static 변수로 할당된 공유 변수에 매번 값을 할당하면서 발생할 수 있는 쓰레드 세이프하지 않은 문제
 - [ ] 예외 처리 시, 오류 메세지 누락으로 인해 디버깅이 어려울 수 있는 문제
 - [ ] 자주 사용하는 Mokcing 대상 객체의 Fake 적용
