+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.Section;
-import nextstep.subway.line.dto.ShortestPathResponse;
+import nextstep.subway.path.dto.ShortestPathResponse;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.dto.StationResponse;
 import org.jgrapht.GraphPath;
